requirements.txt
anyio==3.5.0
base58==2.1.1
black==22.1.0
blinker==1.5
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
click==8.0.3
cryptography==37.0.4
Deprecated==1.2.13
ECPy==1.2.5
Flask==2.0.2
Flask-Login==0.6.2
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pycparser==2.21
pyOpenSSL==22.0.0
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
tomli==2.0.0
types-Deprecated==1.2.9
typing_extensions==4.4.0
urllib3==1.26.8
websockets==10.1
Werkzeug==2.0.2
wrapt==1.14.1
xrpl-py==1.7.0
xrplpers==0.0.4
# git+https://github.com/audiotarky/xrplpers@master#egg=xrplpers
